fix: create useActives hook and fix typo
